Renamed folder from "includes" to "inc"
