+++ ini-rw-test/inc/InputFns.hpp
